+++ darksite/cms/test/conftest.py
@@ -13,6 +13,16 @@ class InfoPanelFactory(factory.DjangoModelFactory):
         model = 'cms.InfoPanel'
 
 
+class MediaResourceFactory(factory.DjangoModelFactory):
+    """
+    Factory for generating ``MediaResource`` instances for testing.
+    """
+    youtube_id = factory.Sequence(lambda n: str(n))
+
+    class Meta:
+        model = 'cms.MediaResource'
+
+
 @pytest.fixture
 def info_panel_factory(db):
     """
@@ -20,3 +30,12 @@ def info_panel_factory(db):
         The factory class used to generate info panels for testing.
     """
     return InfoPanelFactory
+
+
+@pytest.fixture
+def media_resource_factory(db):
+    """
+    Returns:
+        The factory class used to generate media resources for testing.
+    """
+    return MediaResourceFactory