+++ CHANGELOG.md
@@ -7,5 +7,5 @@
 * Added user accounts.
 * Added blog post model.
 * Added endpoint to fetch post list.
-* Added media resource model.
-* Added model to track content of information panels shown on home screen.
+* Added media resource model and capability to query for a media resource by ID.
+* Added model to track content of information panels shown on home screen and ability to query for information panels.